Avoid breaking the build on stricter build systems
Some builds, like Linux distro ones, may append -Wall to CFLAGS, causing
the build to break on any warning newer compilers' static analyzers
might include.

In some cases, like this one, the analyzer is not smart enough to verify
that msg is not user-controllable and thus not exploitable. Thus this
change is enough to make the compiler happy in any case.